SPU buffering: use ring-buffer instead of stack
For all buffers a big chunk of memory wass allocated, but only
the bottom part was used (as SPU data is not so big most of
the time.). SPU data was moved arround in a stack wise manner
using memcpy.
Now each SPU buffer has its own memory and the are used as
ring buffer. This save a lot memcpy work and around .5MB less
memory is allocated.