Fix for forced highlight request
When SPU for highlight is activated a forced request for a
highlight event was made. But highlight events can precede
SPU activation. Now the higlight event request is only made
if no highlight has been received yet.

A second issue is that a still frame with indefinite wait
time could come in between. Then userinput is awaited and
the requested highlight will never arrive. The ddvd_wait_highlight
flag is introduced to hold the wait_for_user flag.